+++ app/Http/Controllers/Admin/SizeController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Sizes\IndexSizesRequest;
 use App\Http\Requests\Admin\Sizes\StoreSizeRequest;
 use App\Http\Requests\Admin\Sizes\UpdateSizeRequest;
+use App\Models\Size;
 use App\Services\Admin\Sizes\AdminSizeQueryService;
 use App\Services\Admin\Sizes\AdminSizeService;
 use RuntimeException;
@@ -15,26 +16,33 @@ class SizeController extends Controller
     public function __construct(
         protected AdminSizeQueryService $queries,
         protected AdminSizeService $sizes
-    ) {
-    }
+    ) {}
 
     public function index(IndexSizesRequest $request)
     {
+        $this->authorize('viewAny', Size::class);
+
         return view('admin.size.index', $this->queries->indexData($request->keyword()));
     }
 
     public function trash(IndexSizesRequest $request)
     {
+        $this->authorize('viewAny', Size::class);
+
         return view('admin.size.trash', $this->queries->trashData($request->keyword()));
     }
 
     public function create()
     {
+        $this->authorize('create', Size::class);
+
         return view('admin.size.create');
     }
 
     public function store(StoreSizeRequest $request)
     {
+        $this->authorize('create', Size::class);
+
         $warning = $this->sizes->create($request->validated());
 
         if ($warning) {
@@ -46,16 +54,22 @@ class SizeController extends Controller
 
     public function show($id)
     {
+        $this->authorize('viewAny', Size::class);
+
         return view('admin.size.detail', ['size' => $this->sizes->find((int) $id)]);
     }
 
     public function edit($id)
     {
+        $this->authorize('updateAny', Size::class);
+
         return view('admin.size.edit', ['size' => $this->sizes->find((int) $id)]);
     }
 
     public function update(UpdateSizeRequest $request, $id)
     {
+        $this->authorize('updateAny', Size::class);
+
         $warning = $this->sizes->update((int) $id, $request->validated());
 
         if ($warning) {
@@ -67,6 +81,8 @@ class SizeController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('deleteAny', Size::class);
+
         $this->sizes->softDelete((int) $id);
 
         return redirect()->route('listSize.list')->with('success', 'Đã chuyển size vào thùng rác');
@@ -74,6 +90,8 @@ class SizeController extends Controller
 
     public function restore($id)
     {
+        $this->authorize('restore', Size::class);
+
         try {
             $this->sizes->restore((int) $id);
 
@@ -85,6 +103,8 @@ class SizeController extends Controller
 
     public function forceDelete($id)
     {
+        $this->authorize('forceDelete', Size::class);
+
         try {
             $this->sizes->forceDelete((int) $id);
 
@@ -96,6 +116,8 @@ class SizeController extends Controller
 
     public function search(IndexSizesRequest $request)
     {
+        $this->authorize('viewAny', Size::class);
+
         $route = $request->get('from', 'list') === 'trash' ? 'listSize.trash' : 'listSize.list';
 
         return redirect()->route($route, ['keyword' => $request->keyword()]);
